docs: applicant report and architectural docs cleanup
