+++ frontend/test/features/article_upload/presentation/screens/article_upload_page_test.dart
@@ -11,7 +11,7 @@ import 'package:news_app_clean_architecture/features/article_upload/presentation
 import 'dart:typed_data';
 
 import 'package:news_app_clean_architecture/features/article_upload/domain/entities/article_category.dart';
-import 'package:news_app_clean_architecture/features/article_upload/presentation/pages/article_upload/article_upload_page.dart';
+import 'package:news_app_clean_architecture/features/article_upload/presentation/screens/article_upload/article_upload_page.dart';
 import 'package:news_app_clean_architecture/l10n/generated/app_localizations.dart';
 
 class MockUploadArticleBloc
